FEAT:  Added allBooks State
Created an allBooks state and passed the data via props to BookShelf, from there, all books are passed down via props to listBooks. Lastly, ListBooks loops through all of the books and displays them with the Book Component. This prop will be critical when displaying books by the chosen category.